+++ src/cpp/tests/test_tardigrade_hydra.cpp
@@ -4616,3 +4616,123 @@ BOOST_AUTO_TEST_CASE( test_formMaxRowPreconditioner, * boost::unit_test::toleran
     BOOST_TEST( answer == *hydra.getFlatPreconditioner( ), CHECK_PER_ELEMENT );
 
 }
+
+BOOST_AUTO_TEST_CASE( test_setResidualNorm, * boost::unit_test::tolerance( DEFAULT_TEST_TOLERANCE ) ){
+    /*!
+     * Test setting the norm of the residual
+     */
+
+    floatType time = 1.1;
+
+    floatType deltaTime = 2.2;
+
+    floatType temperature = 5.3;
+
+    floatType previousTemperature = 23.4;
+
+    floatVector deformationGradient = { 0.39293837, -0.42772133, -0.54629709,
+                                        0.10262954,  0.43893794, -0.15378708,
+                                        0.9615284 ,  0.36965948, -0.0381362 };
+
+    floatVector previousDeformationGradient = { -0.21576496, -0.31364397,  0.45809941,
+                                                -0.12285551, -0.88064421, -0.20391149,
+                                                 0.47599081, -0.63501654, -0.64909649 };
+
+    floatVector previousStateVariables = { 0.53155137, 0.53182759, 0.63440096, 0.84943179, 0.72445532,
+                                           0.61102351, 0.72244338, 0.32295891, 0.36178866, 0.22826323,
+                                           0.29371405, 0.63097612, 0.09210494, 0.43370117, 0.43086276,
+                                           0.4936851 , 0.42583029, 0.31226122, 0.42635131, 0.89338916,
+                                           0.94416002, 0.50183668, 0.62395295, 0.1156184 , 0.31728548,
+                                           0.41482621, 0.86630916, 0.25045537, 0.48303426, 0.98555979,
+                                           0.51948512, 0.61289453, 0.12062867, 0.8263408 , 0.60306013,
+                                           0.54506801, 0.34276383, 0.30412079 }; 
+
+    floatVector parameters = { 1, 2, 3, 4, 5 };
+
+    unsigned int numConfigurations = 4;
+
+    unsigned int numNonLinearSolveStateVariables = 5;
+
+    unsigned int dimension = 3;
+
+    class hydraBaseMock : public tardigradeHydra::hydraBase{
+
+        public:
+
+            floatVector jacobian = { -0.15378708,  0.9615284 ,  0.36965948, -0.0381362 , -0.21576496,
+                                     -0.31364397,  0.45809941, -0.12285551, -0.88064421, -0.20391149,
+                                      0.47599081, -0.63501654, -0.64909649,  0.06310275,  0.06365517,
+                                      0.26880192,  0.69886359,  0.44891065,  0.22204702,  0.44488677,
+                                     -0.35408217, -0.27642269, -0.54347354, -0.41257191,  0.26195225 };
+
+            using tardigradeHydra::hydraBase::hydraBase;
+
+            virtual void formNonLinearProblem( ) override{
+
+                floatVector residual( 5, 0 );
+
+                const floatVector *X = getUnknownVector( );
+
+                for ( unsigned int i = 0; i < 5; i++ ){
+                    for ( unsigned int j = 0; j < 5; j++ ){
+                        residual[ i ] += jacobian[5*i+j] * ( *X )[j];
+                    }
+                }
+
+                tardigradeHydra::unit_test::hydraBaseTester::set_residual( *this, residual );
+
+                tardigradeHydra::unit_test::hydraBaseTester::set_flatJacobian( *this, jacobian );
+
+            }
+
+            virtual const unsigned int getNumUnknowns( ) override{ return 5; }
+
+    };
+
+    hydraBaseMock hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                         { }, { },
+                         previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
+
+    floatVector unknownVector = { 0.39293837, -0.42772133, -0.54629709,  0.10262954,  0.43893794 };
+
+    floatType answer = 1.6716509825117496;
+
+    tardigradeHydra::unit_test::hydraBaseTester::set_unknownVector( hydra, unknownVector );
+
+    BOOST_TEST( answer == *hydra.get_residualNorm( ) );
+
+    floatVector dResidualNormdX( 5, 0 );
+
+    floatType eps = 1e-6;
+
+    for ( unsigned int i = 0; i < 5; i++ ){
+
+        floatType delta = eps * std::fabs( unknownVector[ i ] ) + eps;
+
+        floatVector xp = unknownVector;
+
+        floatVector xm = unknownVector;
+
+        xp[ i ] += delta;
+
+        xm[ i ] -= delta;
+
+        hydraBaseMock hydrap( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                              { }, { },
+                              previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
+
+        hydraBaseMock hydram( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                              { }, { },
+                              previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
+
+        tardigradeHydra::unit_test::hydraBaseTester::set_unknownVector( hydrap, xp );
+
+        tardigradeHydra::unit_test::hydraBaseTester::set_unknownVector( hydram, xm );
+
+        dResidualNormdX[ i ] = ( *hydrap.get_residualNorm( ) - *hydram.get_residualNorm( ) ) / ( 2 * delta );
+
+    }
+
+    BOOST_TEST( dResidualNormdX == *hydra.get_dResidualNormdX( ), CHECK_PER_ELEMENT );
+
+}